Start removing LaTeXML from Travis Tests
LaTeXML-Plugin-sTeX used to be a part of the STeX repository and was
tested together with sTeX inside of Travis. Thus it made sense to
configure it before running the tests.

By now, this part has been migrated to a seperate repository and it is
no longer required to run LaTeXML-Plugin-STeX tests within this
repository.

This commit removes the setup code for LaTeXML from .travis.yml.
Progress on the complete removal is tracked in issue #224.